Add LastPost component in SideBar
