--- src/components/LastPost.js
@@ -1,37 +1,0 @@
-import React from "react";
-import { makeStyles } from "@material-ui/styles";
-import Grid from "@material-ui/core/Grid";
-import Paper from "@material-ui/core/Paper";
-import Typography from "@material-ui/core/Typography";
-import ButtonBase from "@material-ui/core/ButtonBase";
-import DraftsIcon from "@material-ui/icons/Drafts";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  paper: {
-    padding: theme.spacing(2),
-    margin: "auto",
-  },
-  title: {
-    padding: 10,
-  },
-  image: {
-    width: 50,
-    height: 50,
-  },
-  img: {
-    margin: "auto",
-    display: "block",
-    maxWidth: "100%",
-    maxHeight: "100%",
-  },
-}));
-
-const LastPost = () => {
-  const classes = useStyles();
-  return <div className={classes.root}></div>;
-};
-
-export default LastPost;
